backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyswisseph==2.10.3.2
sqlalchemy==2.0.31
pydantic==2.7.4
reportlab==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.2
requests==2.32.3
